Bank employees Apis got tested

+++ src/main/java/com/bank/service/EmployeeService.java
@@ -9,6 +9,7 @@ import com.bank.constant.ExceptionConstant;
 import com.bank.entity.Account;
 import com.bank.entity.Address;
 import com.bank.entity.Customer;
+import com.bank.entity.CustomerAccountMapping;
 import com.bank.enums.AccountType;
 import com.bank.exception.BankException;
 import com.bank.repository.AccountRepository;
@@ -18,7 +19,6 @@ import com.bank.repository.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import javax.validation.constraints.NotNull;
 import java.util.*;
 
 @Service
@@ -33,7 +33,7 @@ public class EmployeeService {
     private AccountRepository accountRepository;
 
     @Autowired
-    private CustomerAccoutRepository CustomerAccoutRepository;
+    private CustomerAccoutRepository customerAccoutRepository;
 
     public String addCustomer(CustomerPojo customerPojo) {
         AddressPojo addressPojo = customerPojo.getAddressPojo();
@@ -43,57 +43,106 @@ public class EmployeeService {
         addressRepository.save(address);
         Customer customer = new Customer.Builder().aadhaar(customerPojo.getAadhaar())
                 .email(customerPojo.getEmail()).mobileNo(customerPojo.getMobileNo())
-                .name(customerPojo.getName()).build();
+                .name(customerPojo.getName()).addressId(address.getAddressId()).build();
         customerRepository.save(customer);
-        return BankConstant.ADD_SUCCESS;
+        return BankConstant.CUSTOMER_ADD_SUCCESS;
     }
 
     public String createAccount(AccountRequest accountRequest) {
         AddressPojo addressPojo = accountRequest.getAddressPojo();
 
-        Address address = new Address.Builder().location(addressPojo.getLocation())
-                .city(addressPojo.getCity()).country(addressPojo.getCountry())
-                .pinCode(addressPojo.getPinCode()).state(addressPojo.getState()).build();
-        addressRepository.save(address);
-        Customer customer = new Customer.Builder().aadhaar(accountRequest.getAadhaar())
-                .email(accountRequest.getEmail()).mobileNo(accountRequest.getMobileNo())
-                .name(accountRequest.getName()).build();
-        customerRepository.save(customer);
+//        Address address = new Address.Builder().location(addressPojo.getLocation())
+//                .city(addressPojo.getCity()).country(addressPojo.getCountry())
+//                .pinCode(addressPojo.getPinCode()).state(addressPojo.getState()).build();
+//        addressRepository.save(address);
+//        Customer customer = new Customer.Builder().aadhaar(accountRequest.getAadhaar())
+//                .email(accountRequest.getEmail()).mobileNo(accountRequest.getMobileNo())
+//                .name(accountRequest.getName()).addressId(address.getAddressId()).build();
+//        customerRepository.save(customer);
 
+//        Account account = new Account.Builder().accountNo(accountRequest.getAccountNo()).
+//                accountType(accountRequest.getAccountType().getName()).amount(accountRequest.getAmount())
+//                .customerId(customer.getCustomerId()).build();
         Account account = new Account.Builder().accountNo(accountRequest.getAccountNo()).
                 accountType(accountRequest.getAccountType().getName()).amount(accountRequest.getAmount())
-                .customerId("7").build();
+                .build();
         accountRepository.save(account);
 
         return BankConstant.ACCOUNT_CREATION_SUCCESS;
     }
 
-    public String linkCustomerWithAccounts( String cId, List<AccountRequest> accountRequestList) {
-
-        return null;
-    }
-
-    public String updateKycForCustomer(String cId, CustomerPojo customerPojo) {
-
-
-        return null;
-    }
-
-    public CustomerPojo GetDetailForCustomer(String cId) {
+    public String linkCustomerWithAccounts( int cId, List<String> accountNos) throws BankException {
         Customer customer = customerRepository.findByCustomerId(cId);
-        String addressId = customer.getAddressId();
+        if (customer == null)
+            throw new BankException(ExceptionConstant.CUSTOMER_NOT_EXIST);
+        for (String accountNo : accountNos) {
+            Account account = accountRepository.findByAccountNo(accountNo);
+            if (account != null) {
+//                CustomerAccountMapping customerAccount = new CustomerAccountMapping.Builder()
+//                        .accountNo(accountNo).customerId(cId).build();
+                CustomerAccountMapping customerAccount = customerAccoutRepository.findByAccountId(accountNo);
+                if (customerAccount != null)
+                    customerAccount.setCustomerId(cId);
+                else
+                    customerAccount = new CustomerAccountMapping.Builder().accountNo(accountNo).customerId(cId).build();
+                customerAccoutRepository.save(customerAccount);
+                account.setCustomerId(cId);
+                accountRepository.save(account);
+            }
+        }
+        return BankConstant.ACCOUNT_LINK_SUCCESS;
+    }
+
+
+    public String updateKycForCustomer(int cId, String aadhaar, String mobile, String email) throws BankException {
+        Customer customer = customerRepository.findByCustomerId(cId);
+        if (customer == null)
+            throw new BankException(ExceptionConstant.CUSTOMER_NOT_EXIST);
+        if (aadhaar != null)
+            customer.setAadhaar(aadhaar);
+           // customerRepository.setAadhaar(cId,aadhaar);
+        if (mobile != null)
+            customer.setMobileNo(mobile);
+            //customerRepository.setMobileNo(cId,mobile);
+        if (email != null)
+            customer.setEmail(email);
+           // customerRepository.setEmailId(cId,email);
+        customerRepository.save(customer);
+        return BankConstant.KYC_UPDATE_SUCCESS;
+    }
+
+
+
+    public CustomerPojo GetDetailForCustomer(int cId) throws BankException {
+        Customer customer = customerRepository.findByCustomerId(cId);
+        if (customer == null)
+            throw new BankException(ExceptionConstant.CUSTOMER_NOT_EXIST);
+        int addressId = customer.getAddressId();
         Address address = addressRepository.findByAddressId(addressId);
         AddressPojo addressPojo = AddressPojo.newAddressPojo().city(address.getCity()).
                 country(address.getCountry()).location(address.getLocation()).pinCode(address.getPinCode())
                 .state(address.getState()).build();
-        List<String> accountNoList = CustomerAccoutRepository.findAllAcountNoByCustomerId(cId);
+//        List<String> accountNoList = CustomerAccoutRepository.findAllAcountNoByCustomerId(cId);
+//        List<AccountPojo> accountList = new ArrayList<>();
+//        for (String accountNo : accountNoList) {
+//            Account account = accountRepository.findByAccountNo(accountNo);
+//            AccountPojo accountPojo = new AccountPojo.Builder().accountNo(accountNo).
+//                    accountType(AccountType.valueOf(account.getAccountType())).amount(account.getAmount()).build();
+//            accountList.add(accountPojo);
+//        }
+
+
+        List<CustomerAccountMapping> customerAccounts = customerAccoutRepository.findByCustomerId(cId);
         List<AccountPojo> accountList = new ArrayList<>();
-        for (String accountNo : accountNoList) {
-            Account account = accountRepository.findByAccountNo(accountNo);
-            AccountPojo accountPojo = new AccountPojo.Builder().accountNo(accountNo).
-                    accountType(AccountType.valueOf(account.getAccountType())).amount(account.getAmount()).build();
+        for (CustomerAccountMapping customerAccountMapping : customerAccounts) {
+            Account account = accountRepository.findByAccountNo(customerAccountMapping.getAccountNo());
+            AccountPojo accountPojo = new AccountPojo.Builder().accountNo(customerAccountMapping.getAccountNo()).
+                    accountType(AccountType.valueOf(account.getAccountType().toUpperCase())).amount(account.getAmount()).build();
             accountList.add(accountPojo);
         }
+
+
+
         CustomerPojo customerPojo = new CustomerPojo.CustomerBuilder().setAddressPojo(addressPojo).setAadhaar(customer.getAadhaar())
         .setCid(customer.getCustomerId()).setEmail(customer.getEmail()).setMobileNo(customer.getMobileNo())
         .setName(customer.getName()).accountList(accountList).build();
@@ -102,27 +151,39 @@ public class EmployeeService {
 
 
 
-    public String deleteCustomer(String cId) {
+    public String deleteCustomer(int cId) throws BankException {
+        Customer customer = customerRepository.findByCustomerId(cId);
+        if (customer == null)
+            throw new BankException(ExceptionConstant.CUSTOMER_NOT_EXIST);
         customerRepository.deleteById(cId);
         return BankConstant.DELETE_SUCCESS;
     }
 
-    public int accountBalanceForAccount(String accountNo) {
+    public int accountBalanceForAccount(String accountNo) throws BankException {
         Account account = accountRepository.findByAccountNo(accountNo);
+        if (account == null) {
+            throw new BankException(ExceptionConstant.ACCOUNT_NOT_EXIST);
+        }
         return account.getAmount();
     }
     public String transferMoney(String sourceAccountNo, String targetAccountNo, int amount) throws BankException {
         Account sourceAccount = accountRepository.findByAccountNo(sourceAccountNo);
+        if (sourceAccount == null) {
+            throw new BankException(ExceptionConstant.SOURCE_ACCOUNT_NOT_EXIST);
+        }
         int netAmount = sourceAccount.getAmount();
         if (netAmount < amount) {
             throw new BankException(ExceptionConstant.NOT_SUFFICIENT_AMOUNT);
         }
         Account targetAccount = accountRepository.findByAccountNo(targetAccountNo);
+        if (targetAccount == null) {
+            throw new BankException(ExceptionConstant.TARGET_ACCOUNT_NOT_EXIST);
+        }
         sourceAccount.setAmount(netAmount-amount);
         targetAccount.setAmount(targetAccount.getAmount()+amount);
         accountRepository.save(sourceAccount);
         accountRepository.save(targetAccount);
-        return null;
+        return BankConstant.TRANSFER_MONEY_SUCCESS;
     }
 
     public String accountStatement(String accountNo, Date startTime, Date endTime) {
@@ -130,16 +191,15 @@ public class EmployeeService {
         return null;
     }
 
-    public float calculateInterest(String accountNo, float interestRate) {
+    public float calculateInterest(String accountNo, float interestRate) throws BankException {
         Account account = accountRepository.findByAccountNo(accountNo);
+        if (account == null) {
+            throw new BankException(ExceptionConstant.ACCOUNT_NOT_EXIST);
+        }
         int amount = account.getAmount();
         float interest = (amount  * interestRate * 1 )% 100;
         return interest;
     }
-
-
-
-
 
 
 }
